+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - TT-1727 - Consolidate collaboration models in backwards compatible way.
 - TT-1734 - Allow adding second user as member to a company profile
 - TT-1716 - Company search stop words
+- TT-1727 - Allow changing role of collaborator
 
 ### Fixed bugs
 - No ticket - Upgrade django to 1.11.23 to fix vulnerability